requirements.txt
torch
torchvision
transformers
pillow
pytest
